students/byeonghwi/requirements.txt
Django==4.0.4
django-cors-headers==3.12.0
mysqlclient==2.1.0
bcrypt==3.2.2
pyjwt==2.4.0